Rename NFT stack
