+++ src/screens/HomeStack/NftStack/components/nft-item-details/components/index.ts
