+++ src/screens/HomeStack/NftStack/components/nft-item-details/components/nft-item-details-attributes.tsx
