+++ src/screens/HomeStack/NftStack/components/nft-item-details/index.ts
