+++ src/screens/HomeStack/NftStack/components/nft-item-details/nft-item-details.tsx
