+++ src/screens/HomeStack/NftStack/nft-collection-details.tsx
@@ -3,21 +3,18 @@ import React, {memo, useCallback} from 'react';
 import {NftCollectionDetails} from '@app/components/nft-collection-details';
 import {useTypedNavigation, useTypedRoute} from '@app/hooks';
 import {NftItem} from '@app/models/nft';
-import {
-  NftDetailsStackParamList,
-  NftDetailsStackRoutes,
-} from '@app/route-types';
+import {NftStackParamList, NftStackRoutes} from '@app/route-types';
 
 export const NftCollectionDetailsScreen = memo(() => {
-  const navigation = useTypedNavigation<NftDetailsStackParamList>();
+  const navigation = useTypedNavigation<NftStackParamList>();
   const {params} = useTypedRoute<
-    NftDetailsStackParamList,
-    NftDetailsStackRoutes.NftCollectionDetails
+    NftStackParamList,
+    NftStackRoutes.NftCollectionDetails
   >();
 
   const onPressNftItem = useCallback(
     (item: NftItem) => {
-      navigation.navigate(NftDetailsStackRoutes.NftItemDetails, {item});
+      navigation.navigate(NftStackRoutes.NftItemDetails, {item});
     },
     [navigation],
   );
